Area Select visual working

+++ resources/shaders/Source/Assets.vert
@@ -8,8 +8,8 @@ layout(location = 2) in int   InTransform;
 layout(location = 3) in ivec2 InIPoint1;
 layout(location = 4) in ivec2 InIPoint2;
 layout(location = 5) in vec4  InColorA;
-layout(location = 6) in ivec2 InFPoint1;
-layout(location = 7) in ivec2 InFPoint2;
+layout(location = 6) in vec2 InFPoint1;
+layout(location = 7) in vec2 InFPoint2;
 
 out int   VSIndex;
 out int   VSID;
@@ -17,8 +17,8 @@ out int   VSTransform;
 out ivec2 VSIPoint1;
 out ivec2 VSIPoint2;
 out vec4  VSColorA;
-out ivec2 VSFPoint1;
-out ivec2 VSFPoint2;
+out vec2 VSFPoint1;
+out vec2 VSFPoint2;
 
 void main() {
     VSIndex =     InIndex;
